articular-examples: applied API improvements

+++ articular-examples/src/main/java/articular/example/labs/techdemos/jme/testjaime/components/Jaime.java
@@ -59,6 +59,6 @@ public final class Jaime implements Component {
     @Override
     public Id getId() {
         return new Component.Id((hashCode() >>> 16) ^
-                GameComponents.JAIME.getEntity().getId().intValue());
+                GameComponents.JAIME.getEntity().getId().longValue());
     }
 }
